+++ Adafruit_GFX_dummy_display.h
@@ -35,10 +35,10 @@ class Adafruit_GFX_dummy_display : public Adafruit_GFX
 {
     public:
         Adafruit_GFX_dummy_display(int16_t w, int16_t h, uint8_t zoom = 1); // Constructor
-	    ~Adafruit_GFX_dummy_display();
-	    void display(void);
+        ~Adafruit_GFX_dummy_display();
+        void display(void);
         void clearDisplay(void);
-	    void drawPixel(int16_t x, int16_t y, uint16_t color);
+        void drawPixel(int16_t x, int16_t y, uint16_t color);
         void drawFastVLine(int16_t x, int16_t y, int16_t h, uint16_t color);
         void drawFastHLine(int16_t x, int16_t y, int16_t w, uint16_t color);
         void fillRect(int16_t x, int16_t y, int16_t w, int16_t h, uint16_t color);
@@ -52,7 +52,7 @@ class Adafruit_GFX_dummy_display : public Adafruit_GFX
         const static uint16_t GREEN  = 5;
         const static uint16_t BLUE   = 6;
     private:
-        void initWindowAndRendered(uint8_t w, uint8_t h);
+        void initWindowAndRendered(int16_t w, int16_t h);
         void setColor(uint16_t color);
         SDL_Window *window_;
         SDL_Renderer *renderer_;
